Solved Bug: queryPromiseGet Can't Be use twice, or it will crash
Solved:
Add People function can only be used once. If the queryPromiseGet is invoked after having run addPeople function once, the program will crash and say "queryPromiseGet" is not a function

----
Reason:
Because in addPeople function, line 316, the result of the query insert is mistakenly store into queryPromiseGet, so after that, queryPromiseGet will become the return value of the insert, rather than a function.
Hence, whenever we want to use it like a function, it will crash because it is no longer a function